Fix: Corrige erro de localStorage no SSR e melhorar validação do token JWT.

+++ src/app/services/auth.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@angular/core';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { Credenciais } from '../models/credenciais';
 import { HttpClient } from '@angular/common/http';
 import { API_CONFIG } from '../config/api.config';
 import { JwtHelperService } from '@auth0/angular-jwt';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root',
@@ -12,33 +13,46 @@ export class AuthService {
 
   constructor(
     private http: HttpClient,
-    private jwtService: JwtHelperService) { }
+    private jwtService: JwtHelperService,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) { }
+
   authenticate(creds: Credenciais) {
     return this.http.post(`${API_CONFIG.baseUrl}/login`, creds, {
       observe: 'response',
       responseType: 'text'
-    })
+    });
   }
 
   successfullLogin(authToken: string) {
-    localStorage.setItem('token', authToken);
+    if (isPlatformBrowser(this.platformId)) {
+      const token = authToken.replace('Bearer ', '');
+      localStorage.setItem('token', token);
+    }
   }
 
   isAuthenticated(): boolean {
-    let token = localStorage.getItem('token');
+
+    if (!isPlatformBrowser(this.platformId)) {
+      return false;
+    }
+
+    const token = localStorage.getItem('token');
+
     if (!token) {
       return false;
     }
 
     try {
       return !this.jwtService.isTokenExpired(token);
-    } catch (e) {
+    } catch {
       return false;
     }
   }
 
-   logout() {
-   localStorage.clear();
+  logout() {
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
-
 }
